[ style ] 컴포넌트 쪼개기
